% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBPs}
\alias{plotBPs}
\title{Boxplot by MP for a range of indicators
Figure 3}
\usage{
plotBPs(
  data,
  indicators = unique(data$indicator),
  target = missing,
  limit = missing,
  yminmax = c(0.1, 0.9),
  lowupp = c(0.25, 0.75)
)
}
\description{
Boxplot by MP for a range of indicators
Figure 3
}
